+++ src/main/java/com/example/demo_mall/mallapi/domain/Todo.java
@@ -1,0 +1,25 @@
+package com.example.demo_mall.mallapi.domain;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.time.LocalDate;
+
+@Entity
+@Table(name = "tbl_todo")
+@Getter
+@ToString
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class Todo {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private long tno;
+
+  private String title;
+  private String writer;
+  private Boolean complete;
+  private LocalDate dueDate;
+}
