+++ src/main/java/com/example/demo_mall/sample/member/dto/Customer.java
@@ -1,14 +1,15 @@
-package com.example.demo_mall.member.dto;
+package com.example.demo_mall.sample.member.dto;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
+import jakarta.persistence.*;
 import jakarta.persistence.Id;
 import lombok.Data;
 
 @Data //@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)  //json 변환만 한다면 해당 어노테이션도 사용가능
-public class ReqMemberDto {
-
+@Entity(name = "customers")
+public class Customer {
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private long customerNumber;
   private String customerName;
   private String customerLastName;
   private String customerFirstName;
