+++ src/main/java/com/example/demo_mall/sample/post/dto/Post.java
@@ -1,4 +1,4 @@
-package com.example.demo_mall.post.dto;
+package com.example.demo_mall.sample.post.dto;
 
 import jakarta.persistence.*;
 import lombok.Data;
