+++ src/main/java/com/example/demo_mall/sample/post/repository/PostJpaRepository.java
@@ -1,6 +1,6 @@
-package com.example.demo_mall.post.repository;
+package com.example.demo_mall.sample.post.repository;
 
-import com.example.demo_mall.post.dto.Post;
+import com.example.demo_mall.sample.post.dto.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostJpaRepository extends JpaRepository<Post, Long> {
