+++ src/main/java/com/example/demo_mall/sample/security/CsrfController.java
@@ -1,4 +1,4 @@
-package com.example.demo_mall.security;
+package com.example.demo_mall.sample.security;
 
 import org.springframework.security.web.csrf.CsrfToken;
 import org.springframework.web.bind.annotation.GetMapping;
