+++ src/main/java/com/example/demo_mall/sample/security/SecurityController.java
@@ -1,4 +1,4 @@
-package com.example.demo_mall.security;
+package com.example.demo_mall.sample.security;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
