+++ src/main/resources/mapper/MemberMapper.xml
@@ -1,16 +1,16 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.example.demo_mall.member.repository.MemberMapper">
-    <insert id="insertMember" parameterType="com.example.demo_mall.member.dto.Member">
+<mapper namespace="com.example.demo_mall.sample.member.repository.MemberMapper">
+    <insert id="insertMember" parameterType="com.example.demo_mall.sample.member.dto.Member">
         INSERT INTO member
             (name, birth, mobile, email, password)
         VALUES(#{name}, #{birth}, #{mobile}, #{email}, #{password});
     </insert>
 
-    <select id="getAllMember" resultType="com.example.demo_mall.member.dto.Member">
+    <select id="getAllMember" resultType="com.example.demo_mall.sample.member.dto.Member">
         select *
         from member
     </select>
-    <select id="getMemberById" parameterType="String" resultType="com.example.demo_mall.member.dto.Member">
+    <select id="getMemberById" parameterType="String" resultType="com.example.demo_mall.sample.member.dto.Member">
         select *
         from member
         where member_id = #{memberId}
