+++ src/test/java/com/example/demo_mall/DemoMallApplicationTests.java
@@ -1,9 +1,11 @@
 package com.example.demo_mall;
 
+import lombok.extern.log4j.Log4j;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
+@Log4j
 class DemoMallApplicationTests {
 
     @Test
