+++ src/test/java/com/example/demo_mall/mallapi/repository/TodoRepositoryTest.java
@@ -1,0 +1,12 @@
+package com.example.demo_mall.mallapi.repository;
+
+import lombok.extern.log4j.Log4j;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@SpringBootTest
+@Log4j
+class TodoRepositoryTest {
+
+}
